Project Cleaning
- Added a Shared.h header for common type aliases and defines
- Changed types in most files to use new type aliases
- Added a Development Configuration as an in between Debug and Release
- Moved Libraries and includes common for both project in External at the root of the solution
- Better Math lib WIP

+++ IonEngine/ConvexMeshCollider.cpp
@@ -7,9 +7,9 @@ using namespace IonEngine;
 
 
 ConvexMeshCollider::ConvexMeshCollider(Mesh* mesh, Vec3f scaling) : MeshCollider(mesh) {
-	int stride;
+	u32 stride;
 	float* posPt = (float*) mesh->getAttributePointer(0, stride); // TODO very unsafe, assumes position is at attribute index 0
-	shape = new btConvexHullShape(posPt, mesh->getVertexCount(), stride);
+	shape = new btConvexHullShape(posPt, mesh->getVertexCount(), static_cast<int>(stride));
 	shape->setLocalScaling(btVector3(scaling.x, scaling.y, scaling.z));
 }
