+++ IonEngine/StringUtils.h
@@ -6,9 +6,9 @@
 
 namespace IonEngine {
 	inline void splitString(const std::string& str, char delim, SimpleList<std::string_view>& list) {
-		int currentPos = 0;
+		size_t currentPos = 0;
 		while (currentPos < str.length()) {
-			int delimPos = (int) str.find(delim, currentPos);
+			size_t delimPos = str.find(delim, currentPos);
 			if (delimPos == std::string::npos) break;
 			list.add(std::string_view(str.data() + currentPos, delimPos - currentPos));
 			currentPos = delimPos;
@@ -16,9 +16,9 @@ namespace IonEngine {
 	}
 
 	inline void splitString(const std::string& str, char delim, SimpleList<std::string>& list) {
-		int currentPos = 0;
+		size_t currentPos = 0;
 		while (currentPos < str.length()) {
-			int delimPos = (int) str.find(delim, currentPos);
+			size_t delimPos = str.find(delim, currentPos);
 			if (delimPos == std::string::npos) {
 				list.add(str.substr(currentPos, str.length() - currentPos));
 				break;
